(Engineering) Fixed compiler warnings
* Specifically, in src/Rose

Issue #138

+++ src/Rose/BinaryAnalysis/Architecture/Jvm.C
@@ -514,7 +514,7 @@ switchSuccessors(const SgAsmJvmInstruction* insn, bool &complete) {
                 std::cout << "WARNING: GOTO!!!\n";
             std::cout << "... switchSuccessors (fall through): insert " << va + defOff << std::endl;
 #endif
-            for (int n{3}; n < nOperands; n++) {        // blame bf8a209471
+            for (size_t n{3}; n < nOperands; n++) {
                 if ((ival = isSgAsmIntegerValueExpression(insn->operand(n)))) {
                     retval.insert(va + ival->get_signedValue());
 #if DEBUG_PRINT
